Factor out utility classes for near search.
