+++ searchlib/src/vespa/searchlib/queryeval/CMakeLists.txt
@@ -39,6 +39,7 @@ vespa_add_library(searchlib_queryeval OBJECT
     monitoring_search_iterator.cpp
     multibitvectoriterator.cpp
     multisearch.cpp
+    near_search_utils.cpp
     nearest_neighbor_blueprint.cpp
     nearsearch.cpp
     nns_index_iterator.cpp
